requirements.txt
Flask>=2.3.0
Jinja2>=3.1.2
gunicorn>=20.1.0
psycopg2-binary>=2.9.9
Werkzeug>=2.3.0